Servició Calificación, validaciones, documentación

+++ controllers/comentario.controller.js
@@ -11,7 +11,7 @@ const getAllComentarios = async (req, res) => {
 
 const createComentario = async (req, res) => {
   try {
-    const { comentario_id, ...newComentario } = req.body; // aquí se elimina la propiedad comentario_id del objeto req.body
+    const { comentario_id, ...newComentario } = req.body; 
     const comentarioCreated = await comentarioService.createComentario(newComentario);
     res.status(201).json(comentarioCreated);
   } catch (error) {
@@ -23,6 +23,12 @@ const updateComentario = async (req, res) => {
   try {
     const id = req.params.id;
     const comentario = req.body;
+
+    const existeComentario = await comentarioService.getComentarioById(id);
+    if (!existeComentario) {
+      return res.status(404).json({ error: 'No existe un comentario con el id especificado' });
+    }
+
     const comentarioUpdated = await comentarioService.updateComentario(comentario, id);
     res.status(200).json(comentarioUpdated);
   } catch (error) {
@@ -30,16 +36,22 @@ const updateComentario = async (req, res) => {
   }
 }
 
-
 const deleteComentario = async (req, res) => {
-    try {
-      const id = req.params.id;
-      await comentarioService.deleteComentario(id);
-      res.status(200).json({ message: 'Comentario eliminado correctamente' });
-    } catch (error) {
-      res.status(500).json({ error: 'Error al eliminar el comentario' });
+  try {
+    const id = req.params.id;
+
+    const existeComentario = await comentarioService.getComentarioById(id);
+    if (!existeComentario) {
+      return res.status(404).json({ error: 'No existe un comentario con el id especificado' });
     }
+
+    await comentarioService.deleteComentario(id);
+    res.status(200).json({ message: 'Comentario eliminado correctamente' });
+  } catch (error) {
+    res.status(500).json({ error: 'Error al eliminar el comentario' });
   }
+}
+
   
   module.exports = {
     getAllComentarios,
